+++ packages/local_auth/local_auth_android/android/src/main/java/io/flutter/plugins/localauth/AuthenticationHelperFactory.java
@@ -1,0 +1,20 @@
+package io.flutter.plugins.localauth;
+
+import androidx.annotation.VisibleForTesting;
+import androidx.fragment.app.FragmentActivity;
+import androidx.lifecycle.Lifecycle;
+import io.flutter.plugin.common.MethodCall;
+
+/** Factory class that assists in creating a {@link AuthenticationHelper} instance. */
+class AuthenticationHelperFactory {
+  @VisibleForTesting
+  public static AuthenticationHelper create(
+      Lifecycle lifecycle,
+      FragmentActivity activity,
+      MethodCall call,
+      AuthenticationHelper.AuthCompletionHandler completionHandler,
+      int[] allowedAuthenticators) {
+    return new AuthenticationHelper(
+        lifecycle, activity, call, completionHandler, allowedAuthenticators);
+  }
+}